Fixed: 500 error만 출력하던 문제 해결 및 setCookie.ts 파일 생성
