+++ src/api/postLogin.ts
@@ -1,7 +1,5 @@
-'use server';
-
-import { cookies } from 'next/headers';
 import { BASE_URL } from './api';
+import setCookie from '../lib/setCookie';
 
 interface UserData {
   email: string;
@@ -18,8 +16,6 @@ const postLogin = async (
   email: string,
   password: string,
 ): Promise<CookieData> => {
-  const cookieStore = cookies();
-
   const res = await fetch(`${BASE_URL}/token`, {
     method: 'POST',
     headers: { 'Content-Type': 'application/json' },
@@ -33,41 +29,24 @@ const postLogin = async (
   if (!res.ok) {
     let errorMessage = '';
     switch (res.status) {
+      case 400:
+        errorMessage = '이메일을 입력하세요.';
+        break;
       case 404:
         errorMessage = '존재하지 않거나 비밀번호가 일치하지 않습니다.';
         break;
       default:
-        errorMessage = '500이네요... 404로 되도록 해봐요';
+        errorMessage = '예상치 못한 오류가 발생하였습니다.';
         break;
     }
+
     throw new Error(errorMessage);
   }
 
   const { item } = result;
   const { token, user } = item;
-  const expireTime = 12 * 60 * 60 * 1000;
 
-  cookieStore.set({
-    name: 'token',
-    value: token,
-    httpOnly: true,
-    path: '/',
-    expires: Date.now() + expireTime,
-  });
-  cookieStore.set({
-    name: 'u_id',
-    value: user.item.id,
-    httpOnly: true,
-    path: '/',
-    expires: Date.now() + expireTime,
-  });
-  cookieStore.set({
-    name: 'userType',
-    value: user.item.type,
-    httpOnly: true,
-    path: '/',
-    expires: Date.now() + expireTime,
-  });
+  setCookie({ token, user });
 
   return { token, user };
 };
